requirements.txt
crewai
load_dotenv
langchain
langchain-google-genai
flask